Fixed Cross Origin Problem! :smiley: :lock:
This problem always arises on my development process, prompting me to
disable security on the browser or run a local server, one of which is
an annoyance, the other makes loading my much project slower.

When I take an image and rotate it into a canvas, I must put it back in
an image so that I can render it later on an WorldRenderer,  by putting
back to Image requires using a function that returns a cross origin
error if the browser is running the file statically on the file system.

Turns out canvas's context's drawImage function can render from other
canvases, so I can simply return the canvas element with the rotated
image to avoid the cross origin issue!

Worked perfectly! My project is now fully static!