Neighbor Selector Class :lock:
A new class was added to handle selection of neighbor tiles.

When a tile is set to a value, all neighbors should have their
probabilities updated based on their "new" surroundings: the radius in
which this happens shouldn't be a hardcoded constant.

A large radius implies that the selected model has structures that
depend on neighbors that are far away. This variable must be carefully
because it will drastically change the performance of the world
generator.